CRAYSAT-498: Fix bugs in `HSNBringupWaiter` found in code review
Fix a couple of bugs found in `HSNBringupWaiter` that were found during
code review:

* Do not use `capture_output` argument to `subprocess.run` because it is
  not supported in Python 3.6, which we are targeting. Also fixed in
  `run_ansible_playbook` function.
* Allow regular expression that parses output to find integers with
  multiple digits.

Refactor the code slightly to allow us to test the correct behavior of
the latter bugfix. Add unit tests and improve an existing unit test to
check for the logging of an error when the status script fails.

Test Description:
All unit tests pass and pycodestyle reports no style violations.