CRAYSAT-561: Create a developer guide in markdown format
Convert the small amount of existing developer documentation from rst to
markdown and document the following:

* SAT project structure and how to add a subcommand.
* How to make authenticated requests to APIs behind the API gateway.
* How to do logging in a subcommand.
* How to add configuration options and retrieve configuration values.
* How to create, edit, and build man pages.

Add a link from `CONTRIBUTING.md` to this developer documentation.

This uses markdown so that it can be rendered by Bitbucket Server, which
unfortunately does not support reStructuredText. Both Bitbucket Cloud
and GitHub support Markdown and reStructuredText, so Markdown will
continue to work for this in the future.

Test Description:
Examine the rendered documents and tested the links in Bitbucket Server
web UI.