CRAYSAT-571: Add code to dump state of HSN before shutting down
Add an `HSNStateRecorder` class to dump the state of the high-speed
network according to the fabric controller API before shutting the
system down. Put most of the logic for interacting with the fabric
controller API in the FabricControllerClient so that it can be easily
reused in the waiting logic during boot up.

Add new unit tests for the newly added `FabricControllerClient` methods.

Remove redundant unit tests that were just testing the base
`APIGatewayClient` class's `get`, `post`, `put`, and `delete` methods
under the guise of being specific tests to the `FabricControllerClient`.

Add unit tests for `HSNStateRecorder`. Most of the work is done in the
`FabricControllerClient`, so the tests are simple.

Add a new configuration file option to control the maximum number of old
HSN states that are stored on the system. Add documentation about this
new option to the top-level SAT main page, and correct the path
documented in the `max_pod_states` config file option.

Test Description:
All unit tests pass, pycodestyle reports no warnings.

Tested by running the following in a virtual environment on
noname-ncn-w001:

    from sat.cli.bootsys.state_recorder import HSNStateRecorder
    hsn_recorder = HSNStateRecorder()
    hsn_recorder.dump_state()
    state = hsn_recorder.get_stored_state()
    print(state)

Verified that dumped state was correct. Also changed value of config
option bootsys.max_hsn_states and checked that it correctly maintained
that configured number of hsn states.