CRAYSAT-571: Handle StateError in KubernetesPodStatusWaiter
Rename the attribute that stores the pod status from the previous
shutdown attempt to `stored_k8s_pod_status`, and make it a property that
handles a possible `StateError` that can occur when attempting to load
the previous state from the latest file.

Also initialize the `k8s_pod_status` attribute in the `__init__` method
to get rid of a warning about this in pycharm.

Test Description:
Unit tests passing. pycodestyle reports no violations.