CRAYSAT-571: Fix a bug in k8s pending pods log message
Fix a bug in the message that is logged when pending k8s pods exist
after waiting.

Test Description:
Tested by running a modified version of `do_mgmt_boot` on noname where
pods exist that would not go to the "Succeeded" or "Running" phase.
Verified the correct message was logged.