CRAYSAT-571: Allow rows to be tuples in Report.add_rows
The KubernetesPodStatusWaiter was passing a set of tuples to the
`Report.add_rows` method, which was not supported by `convert_row`.
There's no reason it can't support this, so add support for it.

Test Description:
Tested on noname with a modified `do_mgmt_boot` function while some pods
were known to not be in the 'Succeeded' or 'Running' phases.