CRAYSAT-570: Remove --state-check-fail-action
Get rid of the `--state-check-fail-action` option that influences
what happens when we fail to get the state of some nodes in HSM
for a session template.  Instead, always behave as though 'force'
was specified for the action and do the operation on the session
template anyway if we can't determine if the nodes in the session
template need the action performed.

Test Description:
All unit tests pass.