CRAYSAT-570: Make sure NCNs are booted in correct order
Make sure the NCNs are booted in the correct order, as documented in the
Admin Guide, namely masters -> storage -> workers. Wait until each group
of NCNs is reachable via SSH before starting the next group. This will
take a little longer, but there is some reason it is important to boot
the nodes in this order. The ssh reachability check for node readiness
is more strict than the ping that is documented in the Admin Guide, but
it is probably pretty comparable. We can convert this to a ping check
later on if we want to improve the performance slightly.

Test Description:
Not tested yet.