CRAYSAT-596: Suspend/resume hms-discovery and do cabinet power control
Add a suspend and resume of the hms-discovery cronjob in k8s when the
system is being shutdown and booted, respectively. Add a new `Waiter`
subclass that waits for the hms-discovery cronjob to be scheduled again
after it has been resumed.

Note that a new dependency has been added on dateutil package. This is
needed to give `datetime` instances the same timezone as the ones
returned by `croniter` (another new dependency), so that they can be
compared. Both `croniter` and `python-datetime` are added to the
`requirements.txt` and `cray-sat.spec`.

Add a function that performs the suspend of the hms-discovery cronjob
along with a power off of the liquid-cooled (mountain) cabinets. This
function is not currently called anywhere, but it will comprise a
`cabinet-power` stage of the `sat bootsys shutdown` command when we
refactor it into stages.

Allow `HSMClient.get_component_xnames` to omit xnames that have "State":
"Empty", so that CAPMC will not return errors when querying for the
state of those xnames.

Add another function that performs the resume of the hms-discovery
cronjob and then waits for the compute node BMCs (node controllers) in
the liquid-cooled (mountain) cabinets to be powered on according to
CAPMC. This will indicate that they are ready to be booted with BOS.
This function is also not currently called anywhere, but it will
comprise a `cabinet-power` stage of the `sat bootsys boot` command when
we refactor it into stages.

Test Description:
Unit tests and pycodestyle pass.

Tested by calling the HMSDiscoveryCronJob methods directly in a script
on noname with debug-level logging enabled in SAT to verify log messages
and checking status of the hms-discovery cronjob with `kubectl get
cronjobs -n services hms-discovery`.

Tested the `do_cabinets_power_off` and `do_cabinets_power_on` functions
on surly2. Verified that all mountain cabinet chassis were powered off
and the hms-discovery cronjob in k8s was suspended after the power off.
Verified that all mountain cabinet chassis and slots were powered on
after the power on. Verified that I could boot the compute nodes with
BOS after the cabinets were powered off.

Tested install of new `cray-sat` RPM on surly2 using `zypper install` to
verify that the new RPM dependencies are met by packages in the
available repos.