CRAYSAT-570: Adjust timeout options for `sat bootsys`
Refine the timeout command-line options for `sat bootsys`, and allow
them to be overridden by options of the same names in the config file.

Add some timeouts. Notably, split the waiting for NCNs to shutdown
cleanly out into its own configurable timeout since this can take a lot
longer than the timeout for an NCN to reach the desired power state
after an IPMI power operation.

Instead of directly accessing the timeouts from the command-line
arguments for `sat bootsys`, use the `get_config_value` function from
`sat.config`, so that it will look for the value from the config file
and allow for that value to be overridden by the command-line argument
if it is specified.

Fix a bug where a default value was being specified for the timeout CLI
arguments. If a default value is specified here, it will always take
precedence over the value in the config file whether the CLI arg is
specified or not, which is not what we want.

Test Description:
Ran the following locally in a virtualenv:

* `sat bootsys -h`
* `sat bootsys boot -h`
* `sat bootsys shutdown -h`
* `sat bootsys boot --list-stages`
* `sat bootsys shutdown --list-stages`

Also ran `./tools/generate_default_config.py -o sat.toml sat/config.py`
and verified that all the timeouts and their default values were
included in the config file.

Tested that each timeout option can be specified in the config file, and
that value will be used if no corresponding CLI arg is given to override
it. Tested that the CLI arg value will be used to override the value in
the config file if specified. Did this by modifying
`sat.cli.bootsys.main` to just print the args and exit. Then modified
the values in the config file to non-defaults. Iterated over all
possible timeout CLI args and called `sat bootsys boot/shutdown` with
those args and verified that they override the config file value.