CRAYSAT-570: Move BOS template options into subparsers
Move the `--uan-bos-template` and `--cle-bos-template` options into the
subparsers for the `boot` and `shutdown` actions, so that they are
specified after the action, e.g. `sat bootsys boot --cle-bos-template
CLE_BOS_TEMPLATE`.

Also collapse some common code into a function, and remove the redfish
options. I believe the redfish options were originally used as the
username and password for the IPMI commands sent to NCNs, but that was
incorrect. A search of the code base for 'args.' turns up no results for
the redfish options in the `bootsys` subpackage.

Test Description:
Ran some `-h` commands to check that the usage looked good still.