CRAYSAT-581: Add kubectl to sat Dockerfile
Add kubectl to the cray-sat docker image by downloading it with curl.
Obtain the kubernetes version from the CLOUD/node-image-kubernetes git
repo, so that we ensure that the version of kubectl in the cray-sat
docker image matches the version in the NCN kubernetes image.

Test Description:
Tested on pepsi-ncn-w001 with dst created container.

    kubectl config view
    kubectl -n services get pods
    kubectl -n services logs $BOD_POD_ID -c boa
    sat k8s
    sat k8s --co-located-replicas