CRAYSAT-694: Move RPM and Docker image to sat product stream
Change the product in the Jenkinsfiles from "shasta-standard" and
"shasta-premium" to "sat" to put the SAT RPM and Docker image in the
sat product stream. Note that we are really only using the Docker image
in Shasta v1.4, and the RPM is not installed anymore.

Test Description:
Push branch and check sat product stream location in artifactory for
both the sat RPM and Docker image.