CRAYSAT-712: Update CHANGELOG.md and .version
Update the changelog to describe all changes made since the last merge
from integration to master. Increment the major version number because
these changes are not backwards compatible. The config file location has
changed, and we are switching to primarly supporting running ``sat`` in
a container rather than directly on the host OS.

Update the `.version` file that defines the Docker image version to
match the latest version in the changelog.

Test Description:
Inspected the rendered CHANGELOG.md in Bitbucket web UI. Checked RPM on
artifactory for correct version number.