CRAYSAT-431: Write site info to S3
This commit modifies the sat showrev and setrev commands to read
and write from an S3 bucket, optionally falling back on a local file.

If S3 is not available, then the local file will be used to read
and write site information, but if S3 is available then it will
always overwrite the local site information.

* add unit tests for showrev and setrev to test S3 behavior.
* update man pages for showrev and setrev.
* slightly modify error handling logic in showrev and setrev
  when loading YAML files.
* Use yaml_dump for dumping YAML instead of constructing
  YAML using a series of 'write()' calls, so that data
  types are always strings and modify unit tests accordingly.

Test Description:
* Unit tests pass
* Ran style checks
* Ran ``sat showrev`` and ``sat setrev`` on groot
* Compiled and viewed man pages

Review feedback:
* Change 2 spaces after periods to 1
* Reword doc string for input_site_data
* Remove unnecessary string conversion as all values are strings
* Remove warning for missing 'Site name' in site file
* Change 2 spaces after period to 1
* Decapitalize doc string in test_write_site_data_s3