CRAYSAT-748: Replace MISSING with '-' for product table
This commit updates MISSING fields to hyphens when images and
recipes are not present, so that it does not look like an error.

This change also updates the examples in the `sat showrev`
man page.

Note: this change does not include a changelog entry because the
previous 'unreleased' entry for SAT_716 is still applicable.

Test Description:
* I ran `sat showrev` --products and inspected the result.
* I viewed the man pages using `sat-man showrev`