CRAYSAT-756, CRAYSAT-757: Stop containerd and containers
This commit introduces two changes to the platform-services stage
of `sat bootsys`. First (SAT_756), it stops all containers running
on Kubernetes NCNs using `crictl`, and second (SAT_757) it stops the
containerd service on these NCNs using `systemctl`.

This commit adds a new module, platform.py containing the functions
for stopping (and eventually starting) platform services. The
do_platform_stop function is what stops the containers and containerd,
while do_platform_start does nothing yet.

The stop_containers() function is used to stop the containers on the remote
host using SSH, while the RemoteServiceWaiter class is used to remotely
start or stop (in this case, stop) containerd, also using SSH.  It checks
that a service is active or inactive before doing anything, and if no
operation is needed, it will not start or stop the service.

For greater efficiency, this commit also changes the base Waiter class
from which the RemoteServiceWaiter inherits to check if self.completed
was set at the beginning of wait_for_completion (e.g. in
pre_wait_action) to avoid needless waiting.

In stages.py, the platform-services stages of shutdown and boot
are changed to reference do_platform_stop and do_platform_start
in the new platform.py.

This commit also removes a significant amount of pre-shasta1.4 code for
platform service start and stop, including the mgmt_ansible.py module
which called ansible plays to shut down platform services.

Finally, this commit adds a changelog entry.

Test Description:
* Added unit tests for new platform.py module
* Manual testing in vshasta:
** Installed SAT in a virtualenv on a vshasta NCN
** Ran:

       (venv) ncn-w001-42e7fafa:~/sat # python
       >>> from sat.cli.bootsys.platform import stop_containers
       >>> stop_containers('ncn-w002')

** In a separate window ran `crictl ps -q` on ncn-w002 and verified
   that the output was empty.

** To test RemoteServiceWaiter, I ran a similar test:

       (venv) ncn-w001-42e7fafa:~/sat # python
       >>> from sat.cli.bootsys.platform import RemoteServiceWaiter
       >>> r = RemoteServiceWaiter('ncn-w002', 'containerd', 'inactive', 30)
       >>> r.wait_for_completion()

** I then checked that `containerd` was stopped on each of the NCNs.

** To test both together, on the NCN I modified sat.cli.bootsys.util.get_ncns to return
   a hard-coded list of NCN hostnames, then I ran:

       (venv) ncn-w001-42e7fafa:~/sat # sat bootsys shutdown --stage platform-services
       Stopping containers running under containerd on all Kubernetes NCNs using crictl
       Stopping containerd on all Kubernetes NCNs using systemctl

** I then checked that `containerd` was stopped on each of the NCNs.
   Upon restarting it, I checked that the output of `crictl ps -q` was
   empty.