CRAYSAT-737: Remove `hsn-bringup` stage from `sat bootsys boot`
Remove the `hsn-bringup` stage from `sat bootsys boot`. This stage
currently attempts to bring up the Slingshot High-Speed Network (HSN)
using an ansible playbook that has been removed in Shasta V1.4. It also
queries state of the HSN using the old fabric controller API, which has
been completely replaced in Shasta V1.4 by a next-gen fabric controller
API.

Also remove the capturing of HSN state that used to occur during the
`capture-state` stage of `sat bootsys shutdown`. That state capture also
uses the old fabric controller API, so it no longer works in Shasta
V1.4.

The code that is used by the `hsn-bringup` stage and the HSN state
capture has been left here, so that it can be updated in the future to
use the next-gen fabric controller.

Test Description:
Tested that I don't see the `hsn-bringup` stage and tested running `sat
bootsys shutdown --stage capture-state` in a python virtual environment
on loki to verify that stage works and doesn't attempt to communicate
with the old fabric controller API.