CRAYSAT-754: Get management NCN hostnames from hosts file
Get the management non-compute node (NCN)  hostnames from the hosts file
rather than from ansible inventory, which is no longer present on 1.4
systems. Note that podman automatically creates a `/etc/hosts` file
within the container, and it is based on the host's version of these
files.

Keep support for getting the hostnames for one or more specified
subroles, to use the terminology of the Hardware State Manager (HSM).
The supported subroles are "managers", "workers", and "storage", keeping
the same names that were formerly used as Ansible group names.

Now that the first master node (i.e. ncn-m001) is the node with external
network connectivity, we must exclude that node from the nodes that are
powered off and powered on via `sat bootsys`. For now, this exclusion is
hardcoded. I can't think of a better way to do it right now.

Unit tests have been added for this new function.

Test Description:
Unit tests pass. Pycodestyle reports no violations. Tested manually by
executing the new `get_mgmt_ncn_hostnames` function with arguments for
each of the possible subroles individually and all subroles together.

Ran version 3.2.0 of SAT under podman with `sat bash` on loki's ncn-w001
and verified that the `/etc/hosts` file in the container has the same
entries for NCNs that are present on the host.