CRAYSAT-763: Start containerd in do_platform_start
This commit adds the starting of containerd in the do_platform_start
function of platform.py.

This commit also adds the ability to optionally enable a service
in addition to starting it.  Add and modify unit tests for this new
functionality.

This commit also improves printing and logging messages in both
do_platform_start as well as do_platform_stop.

Update copyright notices.

Finally, this commit updates the change log.

Test Description:
* Installed SAT in a virtualenv on a vshasta NCN
* After modifying sat.cli.bootsys.util.get_ncns to return a hardcoded
  list of NCNs, I ran the following, and then checked that containerd
  was started on all the management NCNs:

      (venv) ncn-w001:~/sat # sat bootsys boot --stage
          platform-services