CRAYSAT-755: Add stopping/starting of kubelet to platform-services
Add the stopping and disabling of kubelet to the platform-services stage
of `sat bootsys shutdown`. Add the starting and enabling of kubelet to
the platform-services stage of `sat bootsys boot`.

Refactor the `sat.cli.bootsys.platform` module to reduce duplicated code
for performing an action on a service across a set of hosts in parallel.

Add a prompt for the admin to confirm that the NCNs discovered by `sat
bootsys` are correct before doing a start or stop of platform services.
This is important because if this is wrong, the system will not shutdown
or boot correctly and may be hard to restore to working order.

Also ensure that if a step of the starting or stopping of platform
services fails, e.g. the stopping of kubelet, the whole stage will fail.
Otherwise, we may do things out of order, e.g. stopping containers
before stopping kubelet, which may be difficult to recover from.

Test Description:
Unit tests pass. pycodestyle passes.

Tested on a virtual shasta system with 1 manager node and 3 worker
nodes. First modified the `/etc/hosts` file to contain the hostnames of
the management NCNs and their IP addresses. Validated that the prompt
showing NCN hostnames allowed me to halt the stage by answering no or
continue by entering yes. Verified that running `sat bootsys shutdown
--stage platform-services` resulted in containerd being stopped and
kubelet being stopped and disabled on all K8s nodes. Verified that
running `sat bootsys boot --stage platform-services` resulted in
containerd and kubelet being started and enabled.