CRAYSAT-765: Unfreeze Ceph during boot
This commit adds unfreezing the Ceph cluster to the platform-services
stage of SAT bootsys boot. It also modifies the ceph_healthy function
to accept when an OSDMAP_FLAGS warning is present, but only if the flags
are those which are considered acceptable, i.e. those that amount
to the cluster being frozen.

This changes the behavior of ceph_healthy in the shutdown context too,
which should consider Ceph healthy if it's already been frozen.

This change refactors the ceph_healthy() function to reduce complexity
a little, adding validate_ceph_warning_state() as a function called by
ceph_healthy().

This function renames 'freeze_ceph' to 'toggle_ceph_freeze_flags' and
enhances it to set or unset the flags, instead of just setting them.

Add unit tests and change log entry.

Test Description:
On a vshasta system, I installed this version of SAT in a virtualenv,
I set the 'noout,norecover,nobackfill' flags using the `ceph` command,
then ran `sat bootsys boot --stage platform-services`, checking that
these flags were unset.