CRAYSAT-709: Address Python RSA vulnerability
This pins the version of the Python RSA library to version 4.7, to
resolve a timing attack vulnerability discovered by the Snyk tool.

For a full description of the vulnerability, see:
https://snyk.io/vuln/SNYK-PYTHON-RSA\-1038401

This commit also adds a change log entry detailing that PyYAML
and RSA were both updated to address security vulnerabilities.

Test Description: I pushed this branch and verified the SAT Docker
image could be created. Then, I ran `sat status` on an internal
system to verify it was functional.