CRAYSAT-794: Remove calls to [enable,disable]-dns-conflict-hosts playbooks
Remove the calls to the `enable-dns-conflict-hosts.yml` and
`disable-dns-conflict-hosts.yml` Ansible playbooks. These playbooks no
longer exist in Shasta v1.4. They were used to uncomment (enable) and
comment out (disable) entries in `/etc/hosts` for the management NCNs
that would normally conflict with DNS running in K8s. This is no longer
necessary as the hostnames for the NCNs and their management interfaces
(e.g. ncn-w001 and ncn-w001-mgmt) will always be present in the hosts
file on the NCNs.

Test Description:
Unit tests and pycodestyle pass.