CRAYSAT-794: Remove calls to Ansible playbooks for BGP reset and status
Remove the calls to Ansible playbooks for resetting and checking status
of BGP peering sessions. These playbooks have been removed in Shasta
v1.4, and there is no replacement automation for them. Remove the
`bgp-check` stage of both the shutdown and boot actions because this is
no longer implemented in `sat bootsys`.

Test Description:
Unit tests and pycodestyle pass. Checked that `bgp-check` stage does not
appear in the output of `sat bootsys shutdown --list-stages` or `sat
bootsys boot --list-stages` and checked that it cannot be specified as
the value of the `--stage` option for either action.