CRAYSAT-797: Remove starting and stopping of dhcdp from `sat bootsys`
Remove the call to the `RunningService` context manager that started and
stopped the dhcpd service in systemd from the `ncn-power` stage of `sat
bootsys` because it is no longer necessary. The IP addresses of the NCNs
and their BMCs are now statically set by cloud-init scripts.

Test Description:
pycodestyle and unit tests pass.