CRAYSAT-797: Remove unused `RunningService` context manager
Remove the `RunningService` context manager since it was only used to
start and stop the `dhcpd` service during the `ncn-power` stage of `sat
bootsys`, and that has been removed. It also seems unlikely that we
will want to use this `RunningService` context manager as-is in the
future because we run in a container, and I can't see a use case for
temporarily running services in the sat container. Also remove its unit
tests.

Test Description:
pycodestyle and unit tests pass.