CRAYSAT-657: Implement console monitoring with screen sessions
The `ipmi_console_start.sh` and `ipmi_console_stop.sh` scripts are no
longer provided in Shasta 1.4. In addition, `sat` is now running in a
container and can't launch a process directly on the NCN.

In order to start console logging for NCNs while they are being shut
down or booted in the ``ncn-power`` stage of ``sat bootsys``, use SSH to
connect to ncn-m001 and launch a screen session that invokes `ipmitool
... sol activate` and runs in the background, logging its output to a
log file. To stop console logging, use `ipmitool ... sol deactivate` and
quit all existing screen sessions whose name match a prefix used by SAT.

Keep the implementation of the console logging in a context manager
called `IPMIConsoleLogger`. This context manager starts console logging
upon entry and cleans up upon exit. During boot, the context is kept
open from the moment the NCNs are powered on until they are reachable
via SSH. During shutdown the context is kept open from the moment the
NCNs are shut down until they reach the powered off state according to
ipmitool.

Remove the installation of the `ipmi_console_start.sh` and
`ipmi_console_stop.sh` scripts from `config-docker-sat.sh`.

Test Description:
Tested with a fake version of `ipmitool` on virtual shasta. This fake
version of `ipmitool` supports `deactivate` which returns immediately,
and `activate`, which prints logs to stdout for about a minute and then
waits for input from stdin indefinitely. This is meant to mimic the
behavior of actual `ipmitool` commands. It also prints whether it
detected the presence of the `IPMITOOL_PASSWORD` environment variable.

Tested using a script that enters the `IPMIConsoleLogger` context and
then waits until the user answers a prompt to exit the context manager.
This allows me to look at how console logging is set up inside the
context and how it gets torn down when I exit.

Tested the following scenarios:
* No console logging directory exists yet, no screen sessions exist yet.
  The directory is created, screen sessions are created, and they log to
  files in that directory.
* Screen sessions exist but do not match the SAT prefix. Existing screen
  sessions are left alone. New ones are created.
* Killed the test script with Ctrl-C and verified that the screen
  sessions are left running and the `ipmitool` commands are active.
* Tested by running again when existing screen sessions exist that do
  match the prefix. They were quit, and new ones were created.