CRAYSAT-760: etcd backup and stop during bootsys platform-services
Add steps to the `platform-services` stage of `sat bootsys shutdown` to
create an etcd snapshot on all the Kubernetes manager nodes and stop
etcd on those nodes.

If etcd is not running on an NCN, it will skip saving a snapshot and log
an info message to that effect. This will allow the `platform-services`
stage to be run again in case the first time fails.

Unit tests still needed.

Test Description:
pycodestyle passes. Existing unit tests pass.

Copied to vshasta running with recent enough images to have etcd as a
separate service in systemd. Set up hosts in `/etc/hosts` and then ran
`sat bootsys shutdown --stage platform-services`. Ran it again
afterwards to verify that the check on whether `etcd` is active prior to
saving a snapshot works as expected.