CRAYSAT-760: Add prompt whether to continue if etcd is inactive
Add prompt if the etcd snapshot could not be saved on any hosts because
etcd was already inactive. This will allow the admin to decide if they
really want to continue without creating an etcd backup, which is
potentially risky.

Test Description:
Unit tests pass. pycodestyle passes. Tested on vshasta by running twice
in a row. The first time, etcd was running, the second time it was not.
Tested both ways of answering the prompt to ensure it continued or
aborted appropriately.