CRAYSAT-656: Remove SAT crayctldeploy sub-package
This commit remove the cray-sat-crayctldeploy subpackage from SAT,
which contained an ansible role for installing SAT on NCNs. This is
not used in newer versions, so it is being removed.

This commit also removes the source for this ansible role.

Test Description:
I built the SAT RPM using Jenkins and verified no crayctldeploy
sub-package was created. I also installed the cray-sat RPM on
a virtual shasta system, to verify that the SAT RPM was still
able to be installed.