CRAYSAT-788: Use new Ceph health check in CephHealthWaiter
This commit changes the CephHealthWaiter class to use the more
advanced Ceph check introduced by SAT_758. In order to work better
in a Waiter context, the ceph_healthy function is renamed
check_ceph_health and converted to return None in the successful
case and raise an Exception in the failure case.

Test Description: I installed this version of SAT in a virtualenv
on a vshasta system, and ran:
* sat bootsys boot --stage ceph-check
* sat bootsys shutdown --stage platform-services
* sat bootsys boot --stage platform-services