CRAYSAT-833: Add --bos-templates/bos_templates option
This commit adds the --bos-templates command-line option and the
bos_templates config file option, intended to replace the
--cle-bos-template and --uan-bos-template command-line options and their
respective config file options.

If the new option is not specified, then a warning will be logged,
but then bootsys will default to the previous behavior of using the
values of --cle|uan-bos-template or their respective defaults.

However, this commit also changes the defaults of the deprecated
--cle|uan-bos-template arguments, such that if they are not specified,
then they are omitted. This means it is possible to specify no BOS
templates, in which case a BOSFailure is raised.

Since the default behavior of --cle-bos-template is removed (regex
matching COS templates), the find_cle_bos_template() function is
removed.

Update sat-bootsys and sat man pages accordingly. Minor formatting changes
to the heading in the sat-bootsys man page, and correct a couple typos.

Update change log.

Test Description:
* Added new unit tests
* Ran `sat init` to verify new config option is created.
* Ran `sat bootsys shutdown --stage bos-operations`:
  * With no options specified
  * With --cle-bos-template and --uan-bos-template options
  * With --bos-templates option
  * With all 3 options specified
  * Repeated the above with the config file equivalents