CRAYSAT-836: Update SAT man page
This commit rewords the DESCRIPTION section of the SAT man page to
no longer discuss the BIS server. It also rewords the section describing
the purpose of SAT, to include that it helps with boot/shutdown and
swapping hardware. Finally, a few small typos are fixed.

Test Description: Compiled and viewed the SAT man page