CRAYSAT-785: Improve error handling in `sat showrev`
Handle two Exceptions better:
* ValueError raised by s3.resource function when an invalid value
  is given as the endpoint.
* kubernetes.client.rest.ApiException when the configuration map
  does not exist.

Add unit tests for the error handling, and add change log entry.
Improve doc strings for functions that raise exceptions.

Test Description:
I tested this on a vshasta system, where the config map did not exist,
and tested with an invalid value (i.e. one that did not start with http)
as the configured S3 endpoint.