CRAYSAT-791: Improve error handling of container stop step
Improve the error handling in the step of the `platform-services` stage
of `sat bootsys shutdown` that stops containers running under
containerd.

Split the step that stops containers and stops containerd into two
steps, one that stops the containers, and one that stops containerd.
This way, if the step to stop containers fails to stop containers and
raises a NonFatalPlatformError, and the admin decides to continue, the
code will still stop containerd.

The step that stops containers first checks if containerd is running. If
not, the step is done. Then it checks for running containers with
`crictl ps -q`. If there are none, then the step is done. Then it stops
containers with the same complicated `crictl ps -q | xargs crictl stop`
command. If that command times out, a warning is logged. Finally the
step checks for any remaining running containers and raises a
`NonFatalPlatformError` so the admin can decide if they want to continue
with the shutdown or pause and debug the issue.

Logging is improved to log the container IDs that exist on each node
prior to stopping them and to log which ones are reported as
successfully stopped as well at the debug level.

New unit tests have been added, and existing unit tests modified to test
this new code. Achieved 100% coverage on new `ContainerStopThread`.

Test Description:
Unit tests pass. pycodestyle reports no violations.

Tested on vShasta where I am occasionally able to get the container
stopping to time out. Tested `sat bootsys shutdown --stage
platform-services` in the following cases:

* Containers stop successfully without timeout, and `containerd` is
  stopped.
* Container stopping times out on ncn-m001, and I answer the prompt "no"
  to abort. containerd is not stopped. Run command again, and it succeeds
  and proceeds to stop containerd.
* Container stopping times out, and I answer "yes" to proceed anyway.
  containerd is stopped successfully.