CRAYSAT-507: Add filtering for combination of xname/snapshot
To allow for filtering on the combination of snapshot name
and xname, add an optional xnames paramter to FASClient.get_snapshots
which filters the returned data to only match up with the given
xnames.

This commit also changes the structure of do_firmware in
sat.cli.firmware.main fairly significantly by reducing
the function into several smaller functions.

The name of the snapshots created by `fwclient` are changed
from '<time stamp>-all-xnames' to 'SAT-<timestamp>' because
the former is misleading when some of the snapshots are not
all xnames.

Add more man page examples illustrating that --snapshot
and --xname can be used together.

This commit changes some behaviors related to the handling of
API errors and missing/empty responses.
* When multiple xnames are given and there is an API error
  or invalid xname for just one of them, then an error is
  logged but it is non-fatal, meaning the remaining
  xnames are still queried.
* If the above scenario occurs but it results in no
  firmware being discovered for any of the devices,
  this results in an error.
* Similar to the above scenarios, if one or more snapshot names
  are given (and optionally xnames) which result in no firmware
  being returned, then exit with an error.
* If --xname is given with --snapshots (but no argument
  to --snapshots), log a warning that --xname's value is ignored.
* If one or more xnames is given with one or more snapshot names
  and a given xname is not in a given snapshot, then a warning
  is logged that the xname is not in the snapshot.

The majority of the error-handling logic is moved into the
FASClient class in order to facilitate code organization and
ease of testing. The changes to error handling also address
SAT_584.

Finally, FUS support is removed in this commit. The FUSClient
class is removed, and the _FirmwareClient base class is removed,
making FASClient directly inherit from APIGatewayClient. This
affects the --session-checks stage of `sat bootsys shutdown`
making it only check FAS, not FUS.

Test Description:
* I ran several `sat firmware` commands to test `sat firmware`:
    `sat firmware` without arguments
    `sat firmware` with one --xname argument
    `sat firmware` with multiple --xname arguments
    `sat firmware` with an invalid xname
    `sat firmware --snapshots`
    `sat firmware --snapshots` with a snapshot name
    `sat firmware --snapshots` with muliple snapshots names
    `sat firmware --snapshots --xname` with one xname and with
       several.
* I ran `sat bootsys shutdown --stage session-checks`.