CRAYSAT-853: Use locked dependencies
This commit adds a requirements.lock.txt with the exact versions
of sat dependencies to use when installing SAT. The requirements.txt
file contains less restrictive version pinning, while the
requirements.lock.txt file contains exact versions.

Additionally, this commit adds a requirements-dev.txt file which
is a combination of the former requirements-build.txt and
requirements-test.txt files, but with no version pinning. A
requirements-dev.lock.txt file contains the exact versions of
the dependencies required to build, in addition to the depdendencies
already in requirements.lock.txt. These dependencies are installed
when running unit tests and building man pages.

This commit updates runBuildPrep.sh to install locked dependencies.
Because requirements-dev.lock.txt contains all the install dependencies in
addition to the build/test dependencies, it is a superset of the
install dependencies and only requirements-dev.lock.txt needs to be
installed in runBuildPrep.sh.

The Dockerfile is also updated to use the locked install dependencies,
requirements.lock.txt.

The setup.py file continues to use the less restrictive
requirements.txt for install_requires.

Test Description:
* I ran runBuildPrep.sh locally and then ran unit tests with
  `nosetests`.
* I built the Docker image, and ran `sat --help` in the container
  image.
* Pushed this branch and ran sat-rpm and sat-docker. I then tested
  sat-rpm with the release/shasta\-1.4 branch of the dst shared
  library.