CRAYSAT-882: Raise HMSDiscoveryError when last_schedule_time is None
This changes the code which waits for the hms-discovery cronjob
to handle the case when the reported value of its 'last scheduled'
time is None. Now, the code will return False for this case, resulting
in the waiter not being complete. It also adds a debug-level log
message for this case.

Test Description: Added a unit test for this case.