CRAYSAT-862: Do not power off all compute/application nodes in bootsys
Do not do a hard power off all the compute nodes and application nodes
that remain powered on after the BOS session to shut them down has
completed during the `bos-operations` stage of `sat bootsys shutdown`.
This was originally done because the documentation for the full-system
power off procedure had this as a step. However, this is not a good idea
to include in this stage for multiple reasons:

* Prevents use of this command to shutdown subsets of nodes w/ BOS
* BOS already does issue a hard power off with CAPMC if nodes do not
  shutdown within a configurable graceful shutdown period.
* The full "System Power Off Procedure" already includes steps to power
  off cabinets recursively, and simply powering off a cabinet will also
  do a hard power off of nodes.

Leave the code that does the shutdown of all compute and application
nodes since it may be useful to add to a separate stage that does a hard
power off of compute and/or application nodes. There may be a use case
for this still.

Test Description:
Installed in a python virtualenv on loki. Used this commmand to shut
down a single compute with a BOS session template that specified just
the one compute.