CRAYSAT-636: Fix test_get_latest_next_schedule_time to work in newer Python
Fix the unit test `test_get_latest_next_schedule_time` in
`test_discovery.py` to work in newer versions of python, including
Python 3.8.8, which is the version included in Alpine 3.13.2.

Instead of mocking `issubclass`, make a subclass of `datetime`, override
its `now` class method to return the desired datetime, and then patch in
this version of the `datetime` class. That way, the `issubclass` check
in `croniter` will see the mocked out return type as a subclass of
`datetime.datetime`.

Test Description:
Ran unit tests under both python 3.6.9 (where the original unit test was
written) and python 3.8.8 (the new version targeted in Alpine 3.13.2.