CRAYSAT-708, CRAYSAT-868: Remove ceph-check and re-order Ceph platform steps
This commit changes the behavior of the ``platform-services`` stage
of ``sat bootsys boot``. Previously, this stage would first check
Ceph health and if Ceph was healthy then the cluster would be
un-frozen by un-setting the flags "noout", "nobackfill" and "norecover".

This commit changes the order, so that the stage first un-freezes
Ceph and then waits for Ceph health. Additionally, before un-freezing
Ceph the stage will start any inactive Ceph services.

This commit also slightly changes the logic toward handling OSD flags.
First, there is no longer an option to suppress logging a warning when
"acceptable" OSD flags are set. Second, when starting platform-services
the existence of any OSD flags (even "acceptable" ones) will result
in the health check failing, because at this stage we do not expect
any flags to be set.

This commit includes a minor change to strip() the output from the
commands that freeze/unfreeze Ceph, so that unwanted newlines are not
logged.

This commit also changes the prompt_for_ncn_verification() function
to also prompt for verification of the storage node hosts, because
the platform-services stage now restarts services on the storage nodes.

Finally, the ceph-check stage of `sat bootsys boot` is removed
as restarting Ceph services and waiting for Ceph health are covered
by the `platform-services` stage. As a result, the unused
restart_ceph_services() function is also removed, with restarting
Ceph services being done using do_service_action_on_hosts().

Test Description:
On a Virtual Shasta system, I ran:
* sat bootsys shutdown --stage platform-services
* systemctl stop ceph-mgr.target (on storage node)
* systemctl stop ceph-mon.target (on storage node)
* sat bootsys boot --stage platform-services
* systemctl status ceph-mon.target (on storage node)
* systemctl status ceph-mgr.target (on storage node)