CRAYSAT-948: Fix build of sat man pages in container image
Fix the build of the sat man pages in the container image by installing
the missing `docutils` package before running the `make` command that
calls `rst2man.py` to build the man pages from rst. Use the exact
version of `docutils` specified in `requirements-dev.lock.txt` to ensure
a consistent version is used whenever building the man pages. Uninstall
the `docutils` package from the Docker image when done since it is not
needed in the final product.

Also add a `-e` flag to the `config-docker-sat.sh` script so that it
will exit non-zero when an error is encountered and then fail the build
of the Docker image. This will prevent us from releasing a Docker image
with a problem like this in the future.

Test Description:
Tested a local build of the Docker image first with just the `-e` change
to `config-docker-sat.sh` to ensure that this change causes the build of
the Docker image to fail when the build of man pages in that script
fails. Tested a build with all changes, ran a shell in a container from
the image and verified that sat man pages were present and viewable with
the `man` command, e.g. `man sat` and `man sat-diag`.